Add custom password validator
