--- app/forms/changePassword.py
@@ -1,12 +1,0 @@
-from flask_wtf import FlaskForm
-from wtforms import PasswordField, SubmitField
-from wtforms.validators import DataRequired, EqualTo
-
-class ChangePassword(FlaskForm):
-    current_password = PasswordField("Current Password", validators=[DataRequired()])
-    new_password = PasswordField("New Password", validators=[DataRequired()])
-    confirm_password = PasswordField(
-        "Confirm New Password",
-        validators=[DataRequired(), EqualTo("new_password", message="Passwords must match.")],
-    )
-    submit = SubmitField("Change Password")
